Moved password manager triggers to the class
